File structure of the project was changed to match clean architecture principles. Ui components, ChooseHero.kt and SingleHero.kt were renamed to be named the same as their composable function
