+++ app/src/main/java/com/example/marvel_app_project/MainActivity.kt
@@ -8,7 +8,7 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
 import androidx.core.view.WindowCompat
-import com.example.marvel_app_project.ui.AppNavGraph
+import com.example.marvel_app_project.ui.navigation.AppNavGraph
 import com.example.marvel_app_project.ui.theme.ApplySystemBarColors
 import com.example.marvel_app_project.ui.theme.Marvel_app_projectTheme
 import dagger.hilt.android.AndroidEntryPoint
