+++ app/src/main/java/com/example/marvel_app_project/data/network/Either/EitherCall.kt
@@ -1,4 +1,4 @@
-package com.example.marvel_app_project.network.Either
+package com.example.marvel_app_project.data.network.Either
 
 import com.example.marvel_app_project.models.data.network.ErrorResponse
 import okhttp3.Request
