+++ app/src/main/java/com/example/marvel_app_project/data/network/Either/EitherCallAdapterFactory.kt
@@ -1,4 +1,4 @@
-package com.example.marvel_app_project.network.Either
+package com.example.marvel_app_project.data.network.Either
 
 import com.example.marvel_app_project.models.data.network.ErrorResponse
 import retrofit2.Call
