+++ app/src/main/java/com/example/marvel_app_project/ui/navigation/AppNavGraph.kt
@@ -1,4 +1,4 @@
-package com.example.marvel_app_project.ui
+package com.example.marvel_app_project.ui.navigation
 
 import androidx.compose.runtime.Composable
 import androidx.hilt.navigation.compose.hiltViewModel
@@ -6,10 +6,10 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.example.marvel_app_project.ui.pages.choosehero.ChooseHeroScreen
 import com.example.marvel_app_project.ui.pages.choosehero.ChooseHeroViewModel
-import com.example.marvel_app_project.ui.pages.singlehero.SingleHeroScreen
+import com.example.marvel_app_project.ui.pages.choosehero.screens.ChooseHeroScreen
 import com.example.marvel_app_project.ui.pages.singlehero.SingleHeroViewModel
+import com.example.marvel_app_project.ui.pages.singlehero.screens.SingleHeroScreen
 
 enum class HeroesScreen {
     Start,
@@ -19,17 +19,6 @@ enum class HeroesScreen {
 
 @Composable
 fun AppNavGraph(navController: NavHostController = rememberNavController()){
-
-//    val context = LocalContext.current
-//    val database by lazy {
-//        HeroDatabase.getDatabase(context = context)
-//    }
-//    val repository by lazy {
-//        HeroRepositoryImpl(database.heroDao())
-//    }
-//
-//    val heroViewModel = ChooseHeroViewModel(repository = repository)
-//    val singleHeroViewModel = SingleHeroViewModel(repository = repository)
 
     val heroViewModel = hiltViewModel<ChooseHeroViewModel>()
     val singleHeroViewModel = hiltViewModel<SingleHeroViewModel>()
