+++ app/src/main/java/com/example/marvel_app_project/ui/pages/HeroAction.kt
@@ -1,4 +1,4 @@
-package com.example.marvel_app_project.ui
+package com.example.marvel_app_project.ui.pages
 
 sealed interface HeroAction {
     data class OnHeroImageTapped(val heroId: Int, val heroSeverId: String): HeroAction
