+++ app/src/main/java/com/example/marvel_app_project/ui/pages/singlehero/SingleHeroViewModel.kt
@@ -5,10 +5,10 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.marvel_app_project.domain.HeroRepository
+import com.example.marvel_app_project.data.HeroRepository
+import com.example.marvel_app_project.data.network.Either.Either
 import com.example.marvel_app_project.mappers.toSingleUI
-import com.example.marvel_app_project.network.Either.Either
-import com.example.marvel_app_project.ui.HeroAction
+import com.example.marvel_app_project.ui.pages.HeroAction
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
@@ -30,6 +30,7 @@ class SingleHeroViewModel @Inject constructor(
 
             HeroAction.OnBackToChooseHero ->
                 singleHeroUIState = SingleHeroUiState.Loading
+
         }
     }
 
@@ -47,7 +48,7 @@ class SingleHeroViewModel @Inject constructor(
                     is Either.Success -> SingleHeroUiState.Success(
                         singleHeroUIValue = responseFromRepo.value.toSingleUI()
                     )
-            }
+                }
         }
     }
 }
